+++ src/main/java/dinkplugin/DinkPlugin.java
@@ -58,7 +58,7 @@ import java.util.concurrent.atomic.AtomicReference;
 @Slf4j
 @PluginDescriptor(
     name = "Dink",
-    description = "A notifier for sending webhooks to Discord or other custom destinations",
+    description = "Discord-compatible webhook notifications for Loot, Death, Levels, CLog, KC, Diary, Quests, etc.",
     tags = { "loot", "logger", "collection", "pet", "death", "xp", "level", "notifications", "discord", "speedrun",
         "diary", "combat achievements", "combat task", "barbarian assault", "high level gambles" }
 )