+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## Unreleased
 
+- Minor: Allow customization of region IDs where deaths should be ignored. (#415)
 - Bugfix: Treat deaths in the graveyard room of the mage training arena as safe. (#418)
+- Dev: Clarify the plugin description. (#419)
 
 ## 1.8.3